requirements.txt

cryptography==44.0.2
psycopg2-binary==2.9.9
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
banks==2.1.1
beautifulsoup4==4.13.3
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dataclasses-json==0.6.7
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
docx2txt==0.9
eval_type_backport==0.2.2
fastapi==0.115.12
filetype==1.2.0
frozenlist==1.5.0
fsspec==2025.3.2
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.9.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
greenlet==3.1.1
griffe==1.7.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
llama-cloud==0.1.17
llama-cloud-services==0.6.9
llama-index==0.12.28
llama-index-agent-openai==0.4.6
llama-index-cli==0.4.1
llama-index-core==0.12.28
llama-index-embeddings-google-genai==0.1.0
llama-index-embeddings-mistralai==0.3.0
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.11
llama-index-llms-gemini==0.4.14
llama-index-llms-mistralai==0.4.0
llama-index-llms-openai==0.3.30
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.7
llama-index-readers-llama-parse==0.4.0
llama-index-vector-stores-singlestoredb==0.3.1
llama-parse==0.6.4.post1
lxml==5.3.1
MarkupSafe==3.0.2
marshmallow==3.26.1
mistralai==1.6.0
multidict==6.3.2
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.4
openai==1.70.0
packaging==24.2
pandas==2.2.3
parsimonious==0.10.0
pillow==10.4.0
platformdirs==4.3.7
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.2
pydantic_core==2.33.1
PyJWT==2.10.1
PyMySQL==1.1.1
pyparsing==3.2.3
pypdf==5.4.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9
singlestoredb==0.10.7
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
sqlparams==6.2.0
starlette==0.46.1
striprtf==0.0.26
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
websockets==15.0.1
wrapt==1.17.2
yarl==1.18.3
